requirements.txt
h5py==3.7.0
matplotlib==3.5.2
MyLib==0.0.1
numpy==1.22.3
open3d==0.17.0
opencv_contrib_python==4.6.0.66
opencv_contrib_python_headless==4.6.0.66
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.1.1
Pillow==10.0.0
python_Levenshtein==0.21.0
python_Levenshtein==0.21.1
scikit_learn==1.2.2
scipy==1.8.1
sko==0.5.7
tensorboard==2.10.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.65.0
ultralytics==8.0.119
~==encv_python
~pencv_python==4.6.0.66
